Responsive, fixes (#14)
Responsive
