+++ src/App.scss
@@ -383,6 +383,7 @@ html, body {
     background-size: cover;
     flex: 1;
     position: relative;
+    overflow: hidden;
   }
 }
