+++ src/components/Button/index.scss
@@ -25,6 +25,7 @@
   transition-timing-function: cubic-bezier(0.75, 0, 0.125, 1);
   height: 48px;
   border: 2px;
+  align-items: center;
 
   &.round {
     border-radius: 40px;
