+++ src/screens/Auth/index.scss
@@ -7,6 +7,7 @@
   position: relative;
   display: flex;
   flex-direction: row;
+  overflow: hidden;
   flex: 1;
   z-index: 1;
   &:before {
@@ -145,5 +146,11 @@
     opacity: 1;
     -webkit-transform: translate3d(0, -40%, 0);
     transform: translate3d(0, -40%, 0);
+  }
+}
+
+@media all and (max-width: 780px) {
+  .paper-container {
+    padding: 0;
   }
 }
