+++ src/screens/MyAccount/index.scss
@@ -3,7 +3,7 @@
 .my-account-wrapper {
   display: flex;
   flex: 1;
-  flex-direction: row;
+  flex-direction: column;
   justify-content: flex-start;
   align-self: stretch;
   overflow: hidden;
@@ -18,57 +18,77 @@
     transform: translateX(32%) skew(-42deg);
   }
   & .my-account-menu-container {
-    position: fixed;
-    width: 360px;
+    position: relative;
+    flex: 1;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     height: 100%;
-    z-index: 2;
+    max-height: 60px;
     &:before {
       position: absolute;
       content: "";
-      width: 200%;
-      height: 100%;
-      top: 0;
+      width: 100%;
+      height: 200%;
+      bottom: 0;
       background-color: $ORANGE_BACKGROUND;
-      transform: translate(-30%) skew(5deg);
-      z-index: 1;
+      transform: rotate(-1deg);
     }
 
     & .my-account-menu-wrapper {
-      z-index: 4;
       display: flex;
-      flex-direction: column;
-      align-self: stretch;
+      flex-direction: row;
       justify-content: center;
       position: relative;
       background-color: $PRIMARY_BACKGROUND;
       padding: 12px 0;
       & > div {
-        font-size: 0.86rem;
+        font-size: 0.72rem;
         color: $PRIMARY_TEXT;
         letter-spacing: 2px;
         position: relative;
         cursor: pointer;
         text-transform: uppercase;
-        align-self: flex-start;
-        padding: 16px 64px;
-        width: calc(100% - 45px);
-        z-index: 2;
+        padding: 16px 24px;
+        width: 100%;
         display: block;
+        min-width: 180px;
+        text-align: center;
+
+        &:after {
+          position: absolute;
+          content: '';
+          left: 0;
+          top: 50%;
+          transform: translateY(-50%);
+          width: 1px;
+          height: 60%;
+          background-color: $PRIMARY_BORDER_COLOR;
+        }
+
+        &:before {
+          position: absolute;
+          content: "";
+          left: 50%;
+          bottom: 2px;
+          width: 0;
+          height: 4px;
+          transform: translateX(-50%);
+          background-color: $TERTIARY_BACKGROUND;
+          -webkit-transition: all 250ms ease-in-out;
+          -moz-transition: all 250ms ease-in-out;
+          -o-transition: all 250ms ease-in-out;
+          transition: all 250ms ease-in-out;
+        }
+
         &:hover {
-          &:after {
-            position: absolute;
-            content: "";
-            left: 0;
-            top: 0;
-            width: 100%;
-            height: 100%;
-            background-color: $SECONDARY_BACKGROUND;
-            z-index: -1;
-            display: block;
+          &:before {
+            width: 40%;
+            -webkit-transition: all 250ms ease-in-out;
+            -moz-transition: all 250ms ease-in-out;
+            -o-transition: all 250ms ease-in-out;
+            transition: all 250ms ease-in-out;
           }
         }
 
@@ -76,22 +96,12 @@
           &:before {
             position: absolute;
             content: "";
-            left: 2px;
-            top: 50%;
-            width: 4px;
-            height: 60%;
-            transform: translateY(-50%);
+            left: 50%;
+            bottom: 2px;
+            width: 40%;
+            height: 4px;
+            transform: translateX(-50%);
             background-color: $TERTIARY_BACKGROUND;
-          }
-          &:after {
-            position: absolute;
-            content: "";
-            left: 0;
-            top: 0;
-            width: 100%;
-            height: 100%;
-            z-index: -1;
-            background-color: $SECONDARY_BACKGROUND;
           }
         }
       }
@@ -226,5 +236,18 @@
   100% {
     height: auto;
     opacity: 1;
+  }
+}
+
+@media all and (max-width: 780px) {
+  .my-account-wrapper {
+    & .my-account-content-container {
+      & .my-account-content {
+        padding: 12px;
+        &:before {
+          transform: translateX(0) ;
+        }
+      }
+    }
   }
 }
