+++ src/screens/Product/index.scss
@@ -158,3 +158,40 @@
     opacity: 1;
   }
 }
+
+@media all and (max-width: 780px) {
+  .product-wrapper {
+    flex-direction: column;
+    & .product-image-container {
+      & .product-image-inner-container {
+        max-width: 100vw;
+      }
+    }
+
+    & .product-description-container {
+      & .product-description {
+        max-width: 100%;
+        &:before {
+          transform: translateX(15vw) skew(-35deg);
+        }
+
+        & .product-description-content {
+          max-width: 100%;
+          transform: translateX(0)
+        }
+
+        & .add-to-cart-button-container {
+          transform: translateX(0);
+        }
+
+        & .product-id {
+          transform: translateX(0);
+          align-self: center;
+        }
+      }
+      & .product-sizes-container {
+        transform: translateX(0);
+      }
+    }
+  }
+}
