+++ src/utils/mobile.js
@@ -1,0 +1,7 @@
+const checkIfMobile = () => {
+  return /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
+    navigator.userAgent
+  );
+};
+
+export { checkIfMobile };
